like button, user authent, updated list view

+++ dishes/views.py
@@ -4,6 +4,7 @@ from time import strftime
 
 from django.shortcuts import render, redirect
 from django.core.urlresolvers import reverse
+from django.http import Http404
 from boto.s3.key import Key
 from boto import connect_s3
 
@@ -29,7 +30,7 @@ def detail(request, pk):
 
 def new(request):
     if request.method == "POST":
-        form = DishForm(request.POST, request.FILES)
+        form = DishForm(request.POST, request.FILES, user=request.user)
         if form.is_valid():
             form.save()
 
@@ -39,11 +40,12 @@ def new(request):
     return render(request, 'new.html', {'form': form})
 
 
-
 def edit(request, pk):
     dish = Dish.objects.get(pk=pk)
-    if request.method == "POST":
-        form = DishForm(request.POST, request.FILES, instance=dish)
+    if request.user.username != dish.author.username:
+        raise Http404
+    if (request.method == "POST") and (request.user.username == dish.author.username):
+        form = DishForm(request.POST, request.FILES, user=request.user, instance=dish)
         if form.is_valid():
             d = form.save(commit=False)
             d.save()
@@ -53,14 +55,25 @@ def edit(request, pk):
     return render(request, "edit.html", {"form": form, "dish": dish})
 
 
-
-
 def delete(request, pk):
     dish = Dish.objects.get(pk=pk)
+    if request.user.username != dish.author.username:
+        raise Http404
     if request.method == "POST":
         dish.ingredients.all().delete()
         dish.delete()
         return redirect(reverse('list'))
     return render(request, "delete.html", {"dish": dish})
+
+
+def like(request, pk):
+    if pk:
+        dish = Dish.objects.get(pk=pk)
+        count = dish.likes
+        count += 1
+        dish.likes = count
+        dish.save()
+
+    return redirect(reverse("detail", args=(pk,)))
